Cleaning up
Ordering the chapters a little bit and writing down some notes to start on the chapter 8 exercise after having some sleep.
Probably will give a read over the Java text book and look at the other sources as well as check out the other modules I need for the BSc.